Canvas.setClipDifference use float values, fix #1659

+++ mapsforge-core/src/main/java/org/mapsforge/core/graphics/GraphicContext.java
@@ -69,7 +69,7 @@ public interface GraphicContext {
 
     void setClip(int left, int top, int width, int height, boolean intersect);
 
-    void setClipDifference(int left, int top, int width, int height);
+    void setClipDifference(float left, float top, float width, float height);
 
     void setFilterBitmap(boolean filter);
